Change format of create wallet device entries
Instead of specifying the devices in a dictionary keyed by type, give them
as a list of dictionaries with each dictionary specifying the info for
that device.

For Bitcoin Core wallets, include the RPC url in the device entry

Instead of guessing and providing the Bitcoin Core RPC port, username, and
password, have them explicitly provided via a --rpcurl option. This option
is only used for the watch only wallet. It's information is stored in the
wallet JSON file.